requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
huggingface-hub
sentence-transformers
faiss-cpu
plotly
pandas
numpy
graphviz
